chore(gql): refactor gql errros and separate resolvers
- add new errors
- one resolver for each operation
