+++ packages/challenges/src/core/consts/use-case-error-messages.ts
@@ -4,4 +4,5 @@ export class UseCaseErrorMessages {
   public static readonly TITLE_ALREADY_EXISTS = 'Title already exists';
   public static readonly EMPTY_GITHUB_URL = 'GitHub repository URL cannot be empty';
   public static readonly INVALID_GITHUB_URL = 'Invalid GitHub repository URL';
+  public static readonly INTERNAL_SERVER_ERROR = 'Internal server error';
 }
