+++ packages/challenges/src/infra/http/gql/errors/internal-server-error-gql.error.ts
@@ -1,0 +1,14 @@
+import { UseCaseErrorMessages } from '@/core/consts/use-case-error-messages';
+import { CustomGraphQLError } from './custom-gql.error';
+import { HttpStatus } from '@nestjs/common';
+
+export class InternalServerErrorGraphQLError extends CustomGraphQLError {
+  constructor() {
+    super(UseCaseErrorMessages.INTERNAL_SERVER_ERROR, HttpStatus.INTERNAL_SERVER_ERROR, {
+      code: 'INTERNAL_SERVER_ERROR',
+      statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+    });
+
+    this.stack = undefined;
+  }
+}
